+++ packages/twitch-chat/src/event-handlers/transformBetData.ts
