+++ packages/twitch-chat/src/event-handlers/transformPollData.ts
