+++ packages/twitch-chat/src/index.ts
@@ -4,7 +4,8 @@ import { ApiClient } from '@twurple/api'
 import { use } from 'i18next'
 import FsBackend, { type FsBackendOptions } from 'i18next-fs-backend'
 import { Server } from 'socket.io'
-import { initializeSocket, sendTwitchChatMessage } from './chatClientv2.js'
+import { initializeSocket } from './conduitSetup.js'
+import { sendTwitchChatMessage } from './handleChat.js'
 import { logger } from './logger.js'
 import { getBotAuthProvider } from './twitch/lib/getBotAuthProvider.js'
