+++ packages/twitch-events/src/chatSubIds.ts
@@ -1,5 +1,8 @@
 import type { TwitchEventSubResponse } from './interfaces'
+import type { TwitchEventTypes } from './subscribeChatMessagesForUser'
 
 // Get all existing subscriptions by looping through pages
-
-export const chatSubIds: Record<string, TwitchEventSubResponse['data'][0]> = {}
+export const eventSubMap: Record<
+  string,
+  Record<keyof TwitchEventTypes, Pick<TwitchEventSubResponse['data'][0], 'id' | 'status'>>
+> = {}
